requirements.txt
cycler==0.10.0
decorator==4.3.0
Django==2.1.2
django-widget-tweaks==1.4.3
fa2==0.2
JPype1==0.6.3
kiwisolver==1.0.1
konlpy==0.5.1
matplotlib==3.0.2
mecab-python3==0.7
networkx==2.2
numpy==1.15.4
pandas==0.23.4
Pillow==5.3.0
pyparsing==2.3.0
python-dateutil==2.7.5
python-louvain==0.11
pytz==2018.7
scikit-learn==0.20.1
scipy==1.1.0
six==1.11.0
tqdm==4.28.1
wordcloud==1.5.0